Add aar.
