+++ dConnectDevicePlugin/dConnectDeviceAllJoyn/app/build.gradle
@@ -11,24 +11,23 @@ android {
         versionCode 1
         versionName "1.0"
     }
-    repositories {
-        String MAVEN_URL_BASE = 'https://github.com/DeviceConnect/DeviceConnect-Android/raw/master/'
-        maven {
-            url MAVEN_URL_BASE + 'dConnectDevicePlugin/dConnectDevicePluginSDK/repository'
-        }
-        maven {
-            url MAVEN_URL_BASE + 'dConnectSDK/dConnectSDKForAndroid/repository'
-        }
-    }
+
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    packagingOptions {
+        exclude 'LICENSE.txt'
+        exclude 'META-INF/DEPENDENCIES'
+        exclude 'META-INF/LICENSE'
+        exclude 'META-INF/NOTICE'
+    }
+
 }
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'org.deviceconnect:dconnect-device-plugin-sdk:1.0.0'
+    compile project(':dconnect-device-plugin-sdk')
 }
