--- dConnectDevicePlugin/dConnectDeviceAllJoyn/app/src/main/java/org/deviceconnect/android/profile/LightProfileConstants.java
@@ -1,102 +1,0 @@
-/*
-LightProfileConstants
-Copyright (c) 2014 NTT DOCOMO,INC.
-Released under the MIT license
-http://opensource.org/licenses/mit-license.php
-*/
-
-package org.deviceconnect.android.profile;
-
-import org.deviceconnect.profile.DConnectProfileConstants;
-
-/**
- * ライトプロファイルパラメータ定義.
- * @author NTT DOCOMO, INC.
- */
-public interface LightProfileConstants extends DConnectProfileConstants {
-
-    /**
-     * プロファイル名: {@value} .
-     */
-    String PROFILE_NAME = "light";
-
-    /**
-     * パラメータ: {@value} .
-     */
-    String PARAM_NAME = "name";
-
-    /**
-     * パラメータ: {@value} .
-     */
-    String PARAM_LIGHTS = "lights";
-
-    /**
-     * パラメータ: {@value} .
-     */
-    String PARAM_LIGHT_GROUPS = "lightGroups";
-
-    /**
-     * パラメータ: {@value} .
-     */
-    String PARAM_LIGHT_ID = "lightId";
-
-    /**
-     * パラメータ: {@value} .
-     */
-    String PARAM_ON = "on";
-
-    /**
-     * パラメータ: {@value} .
-     */
-    String PARAM_CONFIG = "config";
-
-    /**
-     * パラメータ: {@value} .
-     */
-    String PARAM_BRIGHTNESS = "brightness";
-
-    /**
-     * パラメータ: {@value} .
-     */
-    String PARAM_COLOR = "color";
-
-    /**
-     * パラメータ: {@value} .
-     */
-    String PARAM_FLASHING = "flashing";
-
-    /**
-     * インターフェース: {@value} .
-     */
-    String INTERFACE_GROUP = "group";
-
-    /**
-     * インターフェース: {@value} .
-     */
-    String ATTRIBUTE_GROUP = "group";
-
-    /**
-     * 属性: {@value} .
-     */
-    String ATTRIBUTE_CREATE = "create";
-
-    /**
-     * 属性: {@value} .
-     */
-    String ATTRIBUTE_CLEAR = "clear";
-
-    /**
-     * パラメータ: {@value} .
-     */
-    String PARAM_GROUP_ID = "groupId";
-
-    /**
-     * パラメータ: {@value} .
-     */
-    String PARAM_LIGHT_IDS = "lightIds";
-
-    /**
-     * パラメータ: {@value} .
-     */
-    String PARAM_GROUP_NAME = "groupName";
-}
