+++ dConnectDevicePlugin/dConnectDevicePluginSDK/repository/org/deviceconnect/dconnect-device-plugin-sdk/1.0.1/dconnect-device-plugin-sdk-1.0.1.pom
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>org.deviceconnect</groupId>
+  <artifactId>dconnect-device-plugin-sdk</artifactId>
+  <version>1.0.1</version>
+  <packaging>aar</packaging>
+  <dependencies>
+    <dependency>
+      <groupId>org.deviceconnect</groupId>
+      <artifactId>dconnect-sdk-for-android</artifactId>
+      <version>1.0.0</version>
+      <scope>compile</scope>
+      <exclusions>
+        <exclusion>
+          <artifactId>dconnect-sdk-for-android</artifactId>
+          <groupId>org.deviceconnect</groupId>
+        </exclusion>
+      </exclusions>
+    </dependency>
+    <dependency>
+      <groupId>dConnectDevicePluginSDK</groupId>
+      <artifactId>dconnect-sdk-for-android</artifactId>
+      <version>unspecified</version>
+      <scope>compile</scope>
+      <exclusions>
+        <exclusion>
+          <artifactId>dconnect-sdk-for-android</artifactId>
+          <groupId>org.deviceconnect</groupId>
+        </exclusion>
+      </exclusions>
+    </dependency>
+  </dependencies>
+</project>
