+++ dConnectDevicePlugin/dConnectDevicePluginSDK/repository/org/deviceconnect/dconnect-device-plugin-sdk/maven-metadata.xml.sha1
@@ -1,1 +1,1 @@
-4e91e30ac5919fe79c9c23bf4d379031b8c33620
+72bafc0f89d6dbeed8b8da5b8592a4557b66cacd
