+++ dConnectDevicePlugin/dConnectDevicePluginSDK/repository/org/deviceconnect/dconnect-sdk-for-android/1.0.1/dconnect-sdk-for-android-1.0.1.pom.sha1
@@ -1,0 +1,1 @@
+b19142af56d6be297f927ed7338b2117876b0af7
