+++ dConnectDevicePlugin/dConnectDevicePluginSDK/repository/org/deviceconnect/dconnect-sdk-for-android/maven-metadata.xml
@@ -1,13 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <metadata>
   <groupId>org.deviceconnect</groupId>
-  <artifactId>dconnect-device-plugin-sdk</artifactId>
-  <version>1.0.0</version>
+  <artifactId>dconnect-sdk-for-android</artifactId>
+  <version>1.0.1</version>
   <versioning>
     <versions>
-      <version>1.0.0</version>
       <version>1.0.1</version>
     </versions>
-    <lastUpdated>20150910090903</lastUpdated>
+    <lastUpdated>20150910090904</lastUpdated>
   </versioning>
 </metadata>
