+++ dConnectDevicePlugin/dConnectDevicePluginSDK/repository/org/deviceconnect/dconnect-sdk-for-android/maven-metadata.xml.sha1
@@ -1,0 +1,1 @@
+bfc9fcbd71f1cd85752785032c1d2aa1075d3e28
