+++ dConnectSDK/dConnectSDKForAndroid/repository/org/deviceconnect/dconnect-sdk-for-android/maven-metadata.xml
@@ -6,7 +6,8 @@
   <versioning>
     <versions>
       <version>1.0.0</version>
+      <version>1.0.1</version>
     </versions>
-    <lastUpdated>20150514060037</lastUpdated>
+    <lastUpdated>20150910090715</lastUpdated>
   </versioning>
 </metadata>
