+++ dConnectSDK/dConnectSDKForAndroid/repository/org/deviceconnect/dconnect-sdk-for-android/maven-metadata.xml.md5
@@ -1,1 +1,1 @@
-fc8b15e6fe5b64ef7804b57e37a13735
+2fff9854ef27d49103c99fb0156ee6c1
