+++ dConnectSDK/dConnectSDKForAndroid/repository/org/deviceconnect/dconnect-sdk-for-android/maven-metadata.xml.sha1
@@ -1,1 +1,1 @@
-151d07cae24c523915d33fd151fa8fd275769195
+4c7021247c088275b1ce010d65aad9677dddb2cd
